🚚(models) restructure tests files for edx models
When statements and events fields have to be tested for the edx models,
two separate files are created for statements and events fields tests
respectively.
